+++ pe/_optimize.py
@@ -182,10 +182,14 @@ def _common_choice(defn):
                     break
                 j += 1
             if j - i > 1:
-                subdefs[i:j] = [Class(ranges)]
+                subdefs[i:j] = [Class(sorted(set(ranges), key=_range_sort_key))]
         i += 1
     return Choice(*subdefs)
 
+
+def _range_sort_key(range: list[tuple[str, str|None]]):
+    """Ensure single hyphen characters are the first."""
+    return (range != ("-", None), range)
 
 def _regex_dot(defn, defs, grpid):
     return Regex('(?s:.)')