+++ pe/_meta.py
@@ -2,4 +2,4 @@
 Meta-information about pe.
 """
 
-__version__ = '0.5.1'
+__version__ = '0.5.2'